+++ CHANGES.md
@@ -1,6 +1,18 @@
+## GRNOC TSDS Services 1.8.2 -- Oct 2024
+
+### Features:
+
+* Adds metadata fields to default install
+
+## GRNOC TSDS Services 1.8.1 -- Oct 2024
+
+### Bug Fixes:
+
+* Fixes tsds_writer queue name loading on startup
+
 ## GRNOC TSDS Services 1.8.0 -- Oct 2024
 
-### Features
+### Features:
 
 * Adds tsds_influx: A Mojolicious based webservice to ingest line protocol / influxdb formated metrics
 * Adds support for environment variable based configuration for container based deployments